requirements.txt
peewee
pymysql
fastapi
strawberry-graphql
load_dotenv
uvicorn